Buffer acquire/release feature is implemented.
Now the provider(slave) can take the buffer of livebox or PD.
and draw something on it directly.

livebox viewer will receive that event but there is no differences
between BUFFER type livebox(or pd) and SCRIPT type livebox(or pd).

So the livebox viewer will care them as same one.

Change-Id: Ib4bb443b3ddcaf54a7a596020fa056f086b41b7d