Add new event path & change pixmap 24 to 32.
Accessibility event paths are added.

Fix the bug of size changed event handler.
Because of invalid fbfile, the size of a box is resetted.
I used instance id instead of fb id,.
So the box buffer info indicates invalid uri
This patch will fix it.

master will send a new fbid when it sending the size,changed event.
Then the viewer will get it and update the fbinfo using received fbid.

Change-Id: I29fc45cc5bd693eb796d4ea256c0b23afe39b192