Use the vconf instead of timer.
To optimize the initiate method.
While using timer, the library tries to make a connection to a master
every N sec later.

But it decrease the system performance, so I just changed the timer
to the vconf notification.

Now the viewer will only tries to connect to the master only if its
state is changed.

But in the worst case,
If the master is started, but the viewer fails to connect to a master,
The viewer will not try to connect to the master again.
Even if the master is ready to accept a new connection.
(If we met this case, the system has serious problems.)

Change-Id: I4bfe71c56fff8a440d12f0655074e12aecf16d33