Remove content file only if destroying the common handler
For the shared view.

If a content file is deleted when a handler of livebox is deleted,
The other handler cannot access to the latest content file.

So the content file should be deleted when the common handler is deleted.
the content file also should be shared with other handlers.

Change-Id: I16e065e6b720e2e1c89b339f87bf58a80d715062