Update "mouse_event == false" type handling.
Previously, when a user release finger, the viewer sends down/move/up event to the provider at once.
So the provider can deal with it as a "clicked" event.

But in case of WebDBox, it requires some gaps for time between press & release touches.

So the viewer should send the mouse down event even though the provider doesn't support the mouse event (mouse_event == false),
And block the sending move event. So the provider will not be able to get the move event while pressing the screen.
But if a user release a finger, the viewer should send the mouse up event.

then we can send the real location of finger for press and release, and elapsed time (up and down).

However we have to care the "ON_HOLD" event for this.

if a provider gets ON_HOLD event, it has not to process the "clicked" event even if a user release a finger from the same area
of pressed area.

[model] Redwood,Kiran,B3(Wearable)
[binary_type] AP
[customer] Docomo/Orange/ATT/Open
[issue#] N/A
[problem]
[cause]
[solution]
[team] HomeTF
[request]
[horizontal_expansion]

Change-Id: I56d19187a363a0e324c296a2538e534476fab89c